fix: Batch 6 — execution/ 去除对 cli/ 的反向依赖（001）
将 parse_limits_override / parse_temperature_argument 从 dayu/cli/arg_parsing.py
下沉到 dayu/execution/cli_execution_options.py（与唯一消费者同模块），消除
execution/ 层对 UI 层的 lazy import 反向依赖。

- dayu/execution/cli_execution_options.py 新增两函数本体 + 本地 _PARSE_MODULE
- dayu/cli/arg_parsing.py 删除原函数体与 _PARSE_MODULE，保留占位注释
- tests/application/test_wechat_main.py 同步切换导入路径
- 无兼容 re-export（遵循 CLAUDE.md「禁止兼容性 re-export」）

验证：pyright 0/0/0；pytest tests/cli tests/application 779 passed。
