+++ dayu/execution/cli_execution_options.py
@@ -10,10 +10,80 @@
 from __future__ import annotations
 
 import argparse
+import json
 from pathlib import Path
 
 from dayu.contracts.toolset_config import ToolsetConfigSnapshot, build_toolset_config_snapshot
-from dayu.execution.options import ExecutionOptions, ExecutionOptionsOverridePayload
+from dayu.execution.options import ExecutionOptions, ExecutionOptionsOverridePayload, normalize_temperature
+from dayu.log import Log
+
+_PARSE_MODULE = "EXECUTION.OPTIONS"
+
+
+def parse_limits_override(
+    raw_json: str | None,
+    *,
+    field_name: str,
+) -> ExecutionOptionsOverridePayload | None:
+    """解析工具 limits JSON 覆盖字符串。
+
+    将 CLI / WeChat 等入口传入的 JSON 字符串解析为
+    ``ExecutionOptionsOverridePayload``，校验值类型（仅允许标量）。
+
+    Args:
+        raw_json: 原始 JSON 字符串；``None`` 表示未提供。
+        field_name: 当前参数名，用于错误提示。
+
+    Returns:
+        归一化后的覆盖字典；未提供时返回 ``None``。
+
+    Raises:
+        SystemExit: 当 JSON 非法、不是对象或包含非标量值时退出。
+    """
+
+    if raw_json is None:
+        return None
+    try:
+        parsed = json.loads(raw_json)
+    except json.JSONDecodeError as exc:
+        Log.error(f"{field_name} 不是合法 JSON: {exc}", module=_PARSE_MODULE)
+        raise SystemExit(2) from exc
+    if not isinstance(parsed, dict):
+        Log.error(f"{field_name} 必须是 JSON 对象", module=_PARSE_MODULE)
+        raise SystemExit(2)
+    normalized: ExecutionOptionsOverridePayload = {}
+    for key, value in parsed.items():
+        if value is None or isinstance(value, str | int | float | bool):
+            normalized[str(key)] = value
+            continue
+        Log.error(f"{field_name} 只允许 JSON 标量值，字段 {key!r} 非法", module=_PARSE_MODULE)
+        raise SystemExit(2)
+    return normalized
+
+
+def parse_temperature_argument(
+    raw_value: str | int | float | None,
+    *,
+    field_name: str,
+) -> float | None:
+    """解析 temperature 参数。
+
+    Args:
+        raw_value: 原始参数值。
+        field_name: 参数名，仅用于错误提示。
+
+    Returns:
+        标准化后的 temperature；未传时返回 ``None``。
+
+    Raises:
+        SystemExit: 当 temperature 非法时退出。
+    """
+
+    try:
+        return normalize_temperature(raw_value, field_name=field_name)
+    except ValueError as exc:
+        Log.error(str(exc), module=_PARSE_MODULE)
+        raise SystemExit(2) from exc
 
 
 def add_execution_option_arguments(parser: argparse.ArgumentParser) -> None:
@@ -107,8 +177,6 @@ def build_execution_options_from_args(args: argparse.Namespace) -> ExecutionOpti
         SystemExit: limits JSON 或 temperature 参数非法时抛出。
     """
 
-    from dayu.cli.arg_parsing import parse_limits_override, parse_temperature_argument
-
     doc_limits = parse_limits_override(
         getattr(args, "doc_limits_json", None),
         field_name="--doc-limits-json",
@@ -186,4 +254,6 @@ def _build_toolset_override_snapshots(
 __all__ = [
     "add_execution_option_arguments",
     "build_execution_options_from_args",
+    "parse_limits_override",
+    "parse_temperature_argument",
 ]
